app/ecommerce_crawler/xianyu_crawler.py
"""咸鱼爬虫"""
from typing import List, Dict, Any
import httpx
from bs4 import BeautifulSoup
from app.ecommerce_crawler.base_crawler import BaseCrawler
from app.utils.logger import get_logger

logger = get_logger()


class XianyuCrawler(BaseCrawler):
    """咸鱼爬虫"""
    
    def __init__(self):
        super().__init__()
        self.base_url = "https://s.2.taobao.com/list/list.htm"
    
    async def search(self, query: str, page: int = 1, page_size: int = 20) -> List[Dict[str, Any]]:
        """搜索商品"""
        try:
            # 注意：咸鱼有较强的反爬虫，这里提供基础实现
            # 生产环境建议使用：
            # 1. Playwright进行真实浏览器渲染
            # 2. Cookie池和代理池
            # 3. 请求频率控制
            
            params = {
                "q": query,
                "cat": "50025969",  # 咸鱼分类
                "s": (page - 1) * page_size
            }
            
            headers = self.get_headers()
            headers.update({
                "Referer": "https://www.taobao.com/",
                "Host": "s.2.taobao.com"
            })
            
            async with httpx.AsyncClient(timeout=self.timeout, follow_redirects=True) as client:
                response = await client.get(self.base_url, params=params, headers=headers)
                
                if response.status_code != 200:
                    logger.warning(f"咸鱼爬虫请求失败: {response.status_code}")
                    return []
                
                # 解析HTML
                soup = BeautifulSoup(response.text, 'lxml')
                
                # 这里需要根据咸鱼实际的HTML结构进行解析
                # 由于咸鱼页面结构经常变化，这里只提供示例
                products = []
                
                # 示例：查找商品列表
                items = soup.find_all('div', class_='item')  # 需要根据实际情况调整
                
                for item in items[:page_size]:
                    product = self._parse_item(item)
                    if product:
                        products.append(product)
                
                logger.info(f"咸鱼爬虫搜索到{len(products)}个商品")
                return products
                
        except Exception as e:
            logger.error(f"咸鱼爬虫失败: {e}")
            return []
    
    async def search_with_playwright(self, query: str, page: int = 1) -> List[Dict[str, Any]]:
        """使用Playwright搜索（更可靠但较慢）"""
        try:
            from playwright.async_api import async_playwright
            
            async with async_playwright() as p:
                browser = await p.chromium.launch(headless=True)
                context = await browser.new_context(
                    user_agent=self.get_random_user_agent(),
                    viewport={'width': 1920, 'height': 1080}
                )
                
                page_obj = await context.new_page()
                
                # 访问搜索页面
                search_url = f"{self.base_url}?q={query}"
                await page_obj.goto(search_url)
                
                # 等待页面加载
                await page_obj.wait_for_load_state('networkidle')
                
                # 提取商品数据
                products = await page_obj.evaluate("""
                    () => {
                        const items = [];
                        // 根据实际页面结构提取数据
                        document.querySelectorAll('.item').forEach(item => {
                            items.push({
                                title: item.querySelector('.title')?.innerText,
                                price: item.querySelector('.price')?.innerText,
                                url: item.querySelector('a')?.href,
                                image: item.querySelector('img')?.src
                            });
                        });
                        return items;
                    }
                """)
                
                await browser.close()
                
                # 标准化数据
                normalized = []
                for product in products:
                    if product.get('title'):
                        normalized.append(self.normalize_product(product, "咸鱼"))
                
                logger.info(f"咸鱼Playwright搜索到{len(normalized)}个商品")
                return normalized
                
        except Exception as e:
            logger.error(f"咸鱼Playwright爬虫失败: {e}")
            return []
    
    def _parse_item(self, item) -> Dict[str, Any]:
        """解析商品项"""
        try:
            # 根据实际HTML结构提取数据
            title_elem = item.find('div', class_='title')
            price_elem = item.find('div', class_='price')
            link_elem = item.find('a')
            image_elem = item.find('img')
            
            if not title_elem or not price_elem:
                return None
            
            product = {
                "platform": "咸鱼",
                "title": title_elem.text.strip(),
                "price": price_elem.text.strip(),
                "original_price": 0,
                "coupon": 0,
                "url": link_elem.get('href', '') if link_elem else '',
                "image": image_elem.get('src', '') if image_elem else '',
                "seller_rating": 0,
                "sales": 0,
            }
            
            return self.normalize_product(product, "咸鱼")
            
        except Exception as e:
            logger.error(f"解析咸鱼商品失败: {e}")
            return None

